Enable biarch support on 64-bit PowerPC
The 64-bit PowerPC toolchain now supports biarch support, which allows
both 32-bit and 64-bit compilation. This currently requires a bit of a
hack in the GNU C library build script because two variants of it need
to be built. Eventually this could be generalized to build multiple
variants for arbitrary multilib toolchains.

Signed-off-by: Thierry Reding <thierry.reding@gmail.com>