Update GCC to 8.2.0
It seems that recent versions of GCC can be built with C++ support in
the bootstrap stage (previously C++ support relied on files that were
only produced in the bootstrap stage of the C library). In fact, C++
support now seems to be required in order to successfully build the
bootstrap stage of the GNU C library.

Signed-off-by: Thierry Reding <thierry.reding@gmail.com>